Block abusive IP addresses and add a new 403 Forbidden page with more info
The pages are in the `html` folder, and the configuration in the two
location configuration files. Note the need for `allow all;` in the block
for the 403 page, else the default 403 page is served because the client
isn't allowed access to the page telling them they are not allowed access!
I renamed the `location-isaac-cs-error` to indicate there were multiple
things in the file now, and moved to using a file like it for physics too.