Add a /images/ path to improve image caching
All content-based images currently have URL paths that look like
    /api/vA.B.C/api/images/path/to/file.svg
which contain the API version number and thus change every time there
is an API release. This limits caching of the images, which are mostly
static on a much longer timescale than our release cycle.
Instead, use this proxy magic to allow the images to use the plain
    /images/path/to/file.jpg
form which will not change. This is safe, since the API serves the
images directly from Git, consistently across released versions.